LayoutTests/http/tests/eventsource/eventsource-bad-mime-type-expected.txt

CONSOLE MESSAGE: EventSource's response has a MIME type ("text/bogus") that is not "text/event-stream". Aborting the connection.
Test EventSource with an event-stream with incorrect mime-type.

PASS
